scraper would respect website guidelines now
added an is_allowed_by_robots function to check if scraping a URL complies with the rules specified in the robots.txt file for that website. This function uses the robotparser module to fetch and parse the robots.txt file. If the URL is not allowed according to the robots.txt rules, the scraper will skip that URL and not make a request.

This enhancement ensures that scraper respects the guidelines set by the robots.txt file and doesn't scrape URLs that are disallowed.